requirements.txt
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
paho-mqtt==2.1.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
